Added scrolling when the current cursor moves off-screen. Fixed window offsetx to use ulongs

+++ src/win/struct.h
@@ -28,7 +28,7 @@ struct win_s
             buffer *b;
             char   *sidebar;
             char   *caption;
-            uint    offsetx, offsety;
+            ulong   offsetx, offsety;
         } leaf;
 
         struct
